+++ moor/lib/src/runtime/query_builder/statements/insert.dart
@@ -93,8 +93,25 @@ class InsertStatement<D extends DataClass> {
   /// This method is used internally by moor. Consider using [insert] instead.
   GenerationContext createContext(Insertable<D> entry, InsertMode mode) {
     _validateIntegrity(entry);
-    final map = table.entityToSql(entry.createCompanion(true))
-      ..removeWhere((_, value) => value == null);
+
+    final rawValues = table.entityToSql(entry.createCompanion(true));
+
+    // strip out null values, apply client defaults where applicable
+    final map = <String, Variable>{};
+    for (final columnName in rawValues.keys) {
+      final value = rawValues[columnName];
+
+      if (value != null) {
+        map[columnName] = value;
+      } else {
+        final column = table.columnsByName[columnName];
+        if (column.defaultValue != null) {
+          map[columnName] = column._evaluateClientDefault();
+        }
+      }
+      // column not set, and doesn't have a client default. So just don't
+      // include this column
+    }
 
     final ctx = GenerationContext.fromDb(database);
     ctx.buffer