+++ moor_generator/lib/src/model/column.dart
@@ -125,6 +125,10 @@ class MoorColumn implements HasDeclaration {
   /// if there is no default expression.
   final String defaultArgument;
 
+  /// Dart code for the `clientDefault` expression, or null if it hasn't been
+  /// set.
+  final String clientDefaultCode;
+
   /// The [UsedTypeConverter], if one has been set on this column.
   final UsedTypeConverter typeConverter;
 
@@ -169,9 +173,10 @@ class MoorColumn implements HasDeclaration {
   /// Whether this column is required for insert statements, meaning that a
   /// non-absent value must be provided for an insert statement to be valid.
   bool get requiredDuringInsert {
+    final hasDefault = defaultArgument != null || clientDefaultCode != null;
     final aliasForPk = type == ColumnType.integer &&
         features.any((f) => f is PrimaryKey || f is AutoIncrement);
-    return !nullable && defaultArgument == null && !aliasForPk;
+    return !nullable && !hasDefault && !aliasForPk;
   }
 
   /// The class inside the moor library that represents the same sql type as
@@ -187,6 +192,7 @@ class MoorColumn implements HasDeclaration {
     this.nullable = false,
     this.features = const [],
     this.defaultArgument,
+    this.clientDefaultCode,
     this.typeConverter,
     this.declaration,
   });